Enhance localization support by adding French and Arabic translations for forum, calendar, and live TV. Update notification links to point to the forum.
