+++ public/locales/ar.json
@@ -12,7 +12,9 @@
         "madinaTv": "تلفزيون المدينة المنورة",
         "ramadan": "تقويم رمضان",
         "namesallah":"99 اسمًا من أسماء الله الحسنى",
-        "forum":"المنتدى"
+        "forum":"المنتدى",
+        "calender":"التقويم",
+        "livetv":"البث التلفزيوني المباشر"
     },
     "banner": {
         "title": "مرحبًا بكم في جلوبال صلاح",
