+++ public/locales/fr.json
@@ -12,7 +12,9 @@
     "madinaTv": "Madina Tv",
     "ramadan": "Calendrier du Ramadan",
     "namesallah":"99 noms d'Allah",
-    "forum":"Forum"
+    "forum":"Forum",
+    "calender":"Calendrier",
+    "livetv":"Télévision en direct"
   },
   "banner": {
     "title": "Bienvenue sur Global Salah",
