+++ src/app/[lang]/question-submitted/page.tsx
@@ -1,72 +1,37 @@
-"use client"
+import React from 'react'
+import QuestionSubmit from './QuestionSubmit'
 
-import { useEffect } from "react"
-import { usePathname, useRouter } from "next/navigation"
-import { Button } from "@/components/ui/button"
-import { Card, CardContent, CardDescription, CardFooter, CardHeader, CardTitle } from "@/components/ui/card"
-import { CheckCircle, ArrowLeft, Home, Info } from "lucide-react"
-import Link from "next/link"
-import { motion } from "framer-motion"
-import { urlSplitter } from "@/lib"
-import { Alert, AlertDescription } from "@/components/ui/alert"
+export async function generateMetadata({ params }: any) {
+  const lang = params.lang
+  return {
+    title: `Question Submitted – Global Salah`,
+    description: ` Your question has been successfully submitted on Global Salah. Our community will review and respond soon. Stay tuned for insightful discussions and answers!`,
+alternates: {
+  canonical: `https://www.globalsalah.com/${lang}/question-submitted`,
+},
+robots: {
+      index: false,
+      follow: false,
+      nocache: true,
+      googleBot: {
+        index: false,
+        follow: false,
+        noimageindex: false,
+        "max-video-preview": -1,
+        "max-image-preview": "large",
+        "max-snippet": -1,
+      },
+    },
+  };
+}
 
-export default function QuestionSubmittedPage() {
-  const router = useRouter()
-  const pathname = usePathname();
-  const lang = urlSplitter(pathname)
 
-  useEffect(() => {
-    const userLoggedIn = localStorage.getItem("isLoggedIn") === "true"
-    if (!userLoggedIn) {
-      router.push(`/${lang}/forum`)
-    }
-  }, [router])
-
+function page() {
   return (
-    <div className="container max-w-md mx-auto py-16 px-4">
-      <motion.div
-        initial={{ opacity: 0, scale: 0.9 }}
-        animate={{ opacity: 1, scale: 1 }}
-        transition={{ duration: 0.5 }}
-      >
-        <Card className="border-primary/20 shadow-md text-center">
-          <CardHeader className="pb-2">
-            <div className="mx-auto w-16 h-16 bg-primary/10 rounded-full flex items-center justify-center mb-4">
-              <CheckCircle className="h-8 w-8 text-emerald-600" />
-            </div>
-            <CardTitle className="text-2xl font-bold">Question Submitted!</CardTitle>
-            <CardDescription className="capitalize">Thank you for contributing to our community</CardDescription>
-          </CardHeader>
-          <CardContent className="pb-6">
-            <p className="text-muted-foreground text-sm mb-4 capitalize">
-              Your question has been submitted and is awaiting approval from our moderators. This process usually takes
-              24-48 hours.
-            </p>
-            <Alert className="bg-muted/50 border-primary/20">
-              <Info className="h-4 w-4" />
-              <AlertDescription className="text-sm text-left">
-              You'll receive a notification on you email once your question is published.
-              </AlertDescription>
-            </Alert>
-   
-          </CardContent>
-          <CardFooter className="flex flex-col gap-2 border-t p-6">
-            <Link href={`/${lang}/forum`} className="w-full">
-              <Button className="w-full gap-2 bg-emerald-500 hover:bg-emerald-700 capitalize text-white">
-                <Home className="h-4 w-4 " />
-                Return to Forum
-              </Button>
-            </Link>
-            <Link href={`/${lang}/add-question`} className="w-full">
-              <Button variant="outline" className="w-full gap-2 border-primary/20 capitalize">
-                <ArrowLeft className="h-4 w-4" />
-                Ask Another Question
-              </Button>
-            </Link>
-          </CardFooter>
-        </Card>
-      </motion.div>
+    <div>
+      <QuestionSubmit/>
     </div>
   )
 }
 
+export default page
